Refactor to remove the need to have Hint type and HintClass
